+++ src/sections/communities/communities.scss
@@ -1,6 +1,6 @@
 @import '../../styles/variables';
 
-.sponsors {
+.communities {
   background-color: $black;
   color: $white;
   padding: 50px 0;
