+++ src/sections/participate/participate.scss
